requirements.txt
matplotlib
requests
openai
dotenv-python